+++ bam_pow/src/api/api_calls.js
@@ -10,23 +10,7 @@ export const getAuthors = () => {
 	})
 }
 
-export const authorCreate = (data) => {
-	console.log('this is the data for create author', data.first_name, data.last_name)
-	return axios({
-		method: 'post',
-		url: apiUrl + '/authors/',
-		data: {
-			author:
-			{first_name: data.first_name,
-			last_name: data.last_name}
-				
-		
-		},
-		headers: {
-			'Content-Type' : 'application/json'
-		},
-	})
-}
+
 export const getIllustrators = () => {
     return axios({
 		method: 'GET',
